requirements.txt
matplotlib
numpy
pandas
streamlit
plotly
wordcloud
arabic-reshaper
python-bidi

